requirements.txt
Flask>=2.0.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.0
joblib>=1.0.0
matplotlib>=3.5.0
Werkzeug>=2.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.0.0
wtforms>=3.0.0
bcrypt>=3.2.0
python-dateutil>=2.8.0